move render and animation methods from Subplot to PlotBase
